+++ src/render/GpuTracingScene.cpp
@@ -35,6 +35,8 @@ namespace {
   static_assert(isKernelRecord<GpuTracingEnvironmentRecord>());
   static_assert(isKernelRecord<GpuTracingDebugIdRecord>());
   static_assert(isKernelRecord<GpuTracingShadingRecord>());
+  static_assert(isKernelRecord<GpuDiffusePathStateRecord>());
+  static_assert(isKernelRecord<GpuDiffusePathStepRecord>());
 
   template<typename Record>
   GpuTracingSceneSectionLayout sectionLayout(GpuTracingSceneSectionKind kind,